pyproject.toml
[project]
name = "areg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.11.14",
    "aiosignal==1.3.2",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "frozenlist==1.5.0",
    "idna==3.10",
    "multidict==6.2.0",
    "openai==1.68.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "ruff==0.11.1",
    "tqdm==4.67.1",
    "urllib3==2.3.0",
    "yarl==1.18.3",
]

[tool.ruff]
ignore = ["E501"]
select = ["E", "F", "B"]
